Add support for cancelling continue() calls
This also removes send/throw from AsyncGenerator, as supporting these would make cancellation after a send or throw unintuitive or impossible. This wasn't a particularly useful feature to be honest, so not a big loss. AsyncGenerator is now just a nice way to make a pipeline directly from a closure without the need for direct invokation, which tends to be a bit clumsy looking to read and write.